# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PR/GR AI Analytics API
  version: 0.4.0
  description: >-
    API модульного MVP: публикации, AI-анализ, решения специалистов, источники,
    синхронный сбор и жизненный цикл НПА.
servers:
  - url: http://127.0.0.1:8000
    description: Local backend
...
paths:
...
  /api/publications:
    get:
      tags: [publications]
      operationId: listPublications
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: source_id
          in: query
          schema: {type: string}
        - name: source_type
          in: query
          schema: {$ref: "#/components/schemas/SourceType"}
        - name: published_from
          in: query
          schema: {type: string, format: date-time}
        - name: published_to
          in: query
          schema: {type: string, format: date-time}
        - name: category
          in: query
          schema: {$ref: "#/components/schemas/Category"}
        - name: proposed_priority
          in: query
          schema: {$ref: "#/components/schemas/Priority"}
        - name: needs_review
          in: query
          schema: {type: boolean}
        - name: visibility
          in: query
          schema:
            type: string
            enum: [active, hidden, all]
            default: active
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Filtered publication feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationList"
    post:
      tags: [publications]
      operationId: createPublication
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PublicationCreate"}
      responses:
        "201":
          description: Manually added publication
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PublicationDetail"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409":
          description: Publication with the same URL or content already exists
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /api/publications/{publication_id}:
    get:
      tags: [publications]
      operationId: getPublication
      parameters:
        - $ref: "#/components/parameters/PublicationId"
      responses:
        "200":
          description: Publication card with latest analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationDetail"
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      tags: [publications]
      operationId: updatePublication
      parameters:
        - $ref: "#/components/parameters/PublicationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PublicationPatch"}
      responses:
        "200":
          description: Publication with a new append-only metadata revision
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PublicationDetail"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationError"}
...
  /api/duplicate-candidates:
    get:
      tags: [deduplication]
      operationId: listDuplicateCandidates
      parameters:
        - name: status
          in: query
          schema:
            allOf:
              - $ref: "#/components/schemas/DuplicateFilterStatus"
            default: unreviewed
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Semantic candidate queue for human review
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DuplicateCandidateList"}
  /api/duplicate-candidates/{candidate_id}/reviews:
    post:
      tags: [deduplication]
      operationId: createDuplicateReview
      parameters:
        - name: candidate_id
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/DuplicateReviewCreate"}
            examples:
              related:
                externalValue: ./examples/duplicate-review.json
      responses:
        "201":
          description: Append-only human verdict and updated candidate projection
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DuplicateCandidate"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationError"}
...
components:
  parameters:
    PublicationId:
      name: publication_id
      in: path
      required: true
      schema: {type: string}
...
    SearchQuery:
      name: q
      in: query
      description: Full-text query over title, content, tags, and entities
      schema: {type: string, minLength: 1}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    Offset:
      name: offset
      in: query
      schema: {type: integer, minimum: 0, default: 0}
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
        details: {type: object, additionalProperties: true}
    SourceType:
      type: string
      enum: [rss, regulator, telegram, telegram_archive, file, seed]
    Category:
      type: string
      enum: [regulation, reputation, competitor, trend, unknown]
    Priority:
      type: string
      enum: [critical, high, medium, low, unknown]
...
    Entity:
      type: object
      additionalProperties: false
      required: [type, value]
      properties:
        type: {type: string}
        value: {type: string}
    Evidence:
      type: object
      additionalProperties: false
      required: [claim, quote]
      properties:
        claim: {type: string}
        quote: {type: string, maxLength: 500}
    Criteria:
      type: object
      additionalProperties: false
      required:
        - business_relevance
        - event_maturity
        - financial_impact
        - implementation_effort
        - risk_severity
        - action_urgency
        - state_support_or_accreditation_change
        - service_or_legal_blocking_risk
        - strategic_technology_status
        - binding_legal_precedent
      properties:
        business_relevance:
          type: [integer, "null"]
          minimum: 0
          maximum: 3
          description: Насколько событие прямо применимо к бизнесу
        event_maturity:
          type: [integer, "null"]
          minimum: 0
          maximum: 3
          description: Насколько событие подтверждено и близко к вступлению в силу
        financial_impact:
          type: [integer, "null"]
          minimum: 0
          maximum: 3
          description: Возможный финансовый эффект
        implementation_effort:
          type: [integer, "null"]
          minimum: 0
          maximum: 3
          description: Сложность необходимых изменений в процессах или продукте
        risk_severity:
          type: [integer, "null"]
          minimum: 0
          maximum: 3
          description: Тяжесть юридического или репутационного риска
        action_urgency:
          type: [integer, "null"]
          minimum: 0
          maximum: 3
          description: Как быстро специалисту нужно отреагировать
        state_support_or_accreditation_change:
          type: boolean
          description: Меняются господдержка, льготы или ИТ-аккредитация
        service_or_legal_blocking_risk:
          type: boolean
          description: Есть прямой риск запрета, блокировки сервиса или уголовной ответственности
        strategic_technology_status:
          type: boolean
          description: Технология или инфраструктура получает стратегический статус
        binding_legal_precedent:
          type: boolean
          description: Есть прямо применимый обязательный правовой или судебный прецедент
    Publication:
      type: object
      additionalProperties: false
      required:
        - id
        - source_id
        - external_id
        - title
        - original_url
        - published_at
        - collected_at
        - content
        - content_hash
        - is_demo
        - latest_analysis_id
        - latest_revision_id
        - tags
        - is_hidden
        - is_manual
        - updated_at
      properties:
        id: {type: string}
        source_id: {type: string}
        external_id: {type: string}
        title: {type: string}
        original_url: {type: string, format: uri}
        published_at: {type: string, format: date-time}
        collected_at: {type: string, format: date-time}
        content: {type: string}
        content_hash: {type: string, pattern: "^sha256:[0-9a-f]{64}$"}
        is_demo: {type: boolean}
        latest_analysis_id: {type: [string, "null"]}
        latest_revision_id: {type: [string, "null"]}
        tags:
          type: array
          items: {type: string, minLength: 1}
        is_hidden: {type: boolean}
        is_manual: {type: boolean}
        updated_at: {type: string, format: date-time}
    PublicationCreate:
      type: object
      additionalProperties: false
      required:
        - source_id
        - title
        - original_url
        - published_at
        - content
        - author_id
      properties:
        source_id: {type: string, minLength: 1}
        title: {type: string, minLength: 1}
        original_url: {type: string, format: uri}
        published_at: {type: string, format: date-time}
        content: {type: string, minLength: 1}
        tags:
          type: array
          uniqueItems: true
          items: {type: string, minLength: 1}
        author_id: {type: string, minLength: 1}
    PublicationPatch:
      type: object
      additionalProperties: false
      required: [author_id]
      properties:
        title: {type: string, minLength: 1}
        tags:
          type: array
          uniqueItems: true
          items: {type: string, minLength: 1}
        is_hidden: {type: boolean}
        author_id: {type: string, minLength: 1}
...
    AnalysisVersion:
      type: object
      additionalProperties: false
      required:
        - id
        - publication_id
        - version
        - analyzer
        - model
        - prompt_version
        - input_hash
        - summary
        - facts
        - entities
        - category
        - proposed_priority
        - criteria
        - importance_score
        - evidence
        - uncertainty
        - needs_review
        - created_at
      properties:
        id: {type: string}
        publication_id: {type: string}
        version: {type: integer, minimum: 1}
        analyzer: {type: string, enum: [replay, live_llm]}
        model: {type: string}
        prompt_version: {type: string}
        input_hash: {type: string, pattern: "^sha256:[0-9a-f]{64}$"}
        summary: {type: string}
        facts: {type: array, items: {type: string}}
        entities: {type: array, items: {$ref: "#/components/schemas/Entity"}}
        category: {$ref: "#/components/schemas/Category"}
        proposed_priority: {$ref: "#/components/schemas/Priority"}
        criteria: {$ref: "#/components/schemas/Criteria"}
        importance_score:
          type: [integer, "null"]
          minimum: 0
          maximum: 18
          description: Детерминированная сумма шести критериев или null при неизвестном значении
        evidence: {type: array, items: {$ref: "#/components/schemas/Evidence"}}
        uncertainty: {type: number, minimum: 0, maximum: 1}
        needs_review: {type: boolean}
        created_at: {type: string, format: date-time}
    PublicationDetail:
      type: object
      additionalProperties: false
      required: [publication, latest_analysis, latest_decision]
      properties:
        publication: {$ref: "#/components/schemas/Publication"}
        latest_analysis:
          oneOf:
            - $ref: "#/components/schemas/AnalysisVersion"
            - type: "null"
        latest_decision:
          oneOf:
            - $ref: "#/components/schemas/SpecialistDecision"
            - type: "null"
    PublicationList:
      type: object
      additionalProperties: false
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/PublicationDetail"}
        total: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1}
        offset: {type: integer, minimum: 0}
    DuplicateStatus:
      type: string
      enum: [unreviewed, duplicate, related, different]
    DuplicateFilterStatus:
      type: string
      enum: [all, unreviewed, duplicate, related, different]
    DuplicateVerdict:
      type: string
      enum: [duplicate, related, different]
    DuplicateReviewCreate:
      type: object
      additionalProperties: false
      required: [verdict, reviewer_id]
      properties:
        verdict: {$ref: "#/components/schemas/DuplicateVerdict"}
        reviewer_id: {type: string, minLength: 1}
        comment: {type: [string, "null"], maxLength: 2000}
    DuplicateReview:
      type: object
      additionalProperties: false
      required: [id, candidate_id, version, verdict, reviewer_id, comment, created_at]
      properties:
        id: {type: string}
        candidate_id: {type: string}
        version: {type: integer, minimum: 1}
        verdict: {$ref: "#/components/schemas/DuplicateVerdict"}
        reviewer_id: {type: string}
        comment: {type: [string, "null"]}
        created_at: {type: string, format: date-time}
    DuplicateCandidate:
      type: object
      additionalProperties: false
      required:
        - id
        - publication
        - candidate_publication
        - model
        - similarity
        - status
        - reviews
        - created_at
      properties:
        id: {type: string}
        publication: {$ref: "#/components/schemas/PublicationDetail"}
        candidate_publication: {$ref: "#/components/schemas/PublicationDetail"}
        model: {type: string}
        similarity: {type: number, minimum: -1, maximum: 1}
        status: {$ref: "#/components/schemas/DuplicateStatus"}
        reviews:
          type: array
          items: {$ref: "#/components/schemas/DuplicateReview"}
        created_at: {type: string, format: date-time}
    DuplicateCandidateList:
      type: object
      additionalProperties: false
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/DuplicateCandidate"}
        total: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1}
        offset: {type: integer, minimum: 0}
...
    DecisionStatus:
      type: string
      enum: [confirmed, corrected, rejected]
...
    SpecialistDecision:
      type: object
      additionalProperties: false
      required:
        - id
        - publication_id
        - analysis_id
        - version
        - status
        - final_summary
        - final_category
        - final_priority
        - comment
        - author_id
        - created_at
      properties:
        id: {type: string}
        publication_id: {type: string}
        analysis_id: {type: string}
        version: {type: integer, minimum: 1}
        status: {$ref: "#/components/schemas/DecisionStatus"}
        final_summary: {type: [string, "null"]}
        final_category: {$ref: "#/components/schemas/Category"}
        final_priority: {$ref: "#/components/schemas/Priority"}
        comment: {type: [string, "null"]}
        author_id: {type: string}
        created_at: {type: string, format: date-time}
...